Precision errors corrected

+++ src/main/java/com/pdelho/lastminute/taxes/model/Receipt.java
@@ -5,8 +5,8 @@ import java.util.List;
 public class Receipt {
 	
 	private List<Product> products;
-	private Double taxes;
-	private Double totalPrice;
+	private Float taxes;
+	private Float totalPrice;
 	
 	
 	public List<Product> getProducts() {
@@ -17,19 +17,19 @@ public class Receipt {
 		this.products = products;
 	}
 	
-	public Double getTaxes() {
+	public Float getTaxes() {
 		return taxes;
 	}
 	
-	public void setTaxes(Double taxes) {
+	public void setTaxes(Float taxes) {
 		this.taxes = taxes;
 	}
 	
-	public Double getTotalPrice() {
+	public Float getTotalPrice() {
 		return totalPrice;
 	}
 	
-	public void setTotalPrice(Double totalPrice) {
+	public void setTotalPrice(Float totalPrice) {
 		this.totalPrice = totalPrice;
 	}
